all new files, delete unused files
